chore: extract shared utilities and add test infrastructure
Extracts escapeHtml, parseBooleanEnv, parseAllowedUsersEnv from
src/index.ts into src/utils.ts for testability.

Adds vitest.config.ts and fast-check devDependency to support
property-based security test suites.
