+++ vitest.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    include: ["src/**/*.test.ts", "packages/**/src/**/*.test.ts", "scripts/**/*.test.ts"],
+  },
+});
